refactor: simplified `Note`
