+++ src/styles/StyledNote.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const Note = styled.div`
+export const StyledNote = styled.div`
   padding: 12px;
   background-color: ${({ theme }) => theme.colors.secondary};
   color: ${({ theme }) => theme.colors.onSecondary};
@@ -8,11 +8,4 @@ export const Note = styled.div`
   display: grid;
   grid-template-columns: auto 1fr;
   align-items: center;
-
-  & div {
-    & > p,
-    & > span {
-      opacity: 0.9;
-    }
-  }
 `;
